requirements.txt
pandas==2.0.3
gspread==5.7.2
python-dotenv==1.0.1
google-auth==2.33.0